+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * Trait: `CallMethod`
 * Helpers: `loadenv`, `env`
 * `e` helper: Add double-encode parameter, pass encoding & flags to `htmlspecialchars`
+* Fix Unwanted Changes in Select
+* Remove `Select::last`
 
 ## [v0.0.02](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.01...v0.0.02) - 8/26/2024