+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Release Notes for 0.x
 
-## [Unreleased](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.02...master)
+## [v0.0.03](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.02...v0.0.03) - 8/29/2024
 
-* Add Response Class
+* Add `Response` Class
 * Implement Service Container
 * Cache facades by getFacadeAccessor
 * Upload Bug Fix